Allow consuming std::array's; allow other output types in consume
This rewrites the tuple/pair code a bit so that it is also usable for
std::array so that things like:

    auto arr = btdp.consume_list<std::array<int, 3>>("v");

now works.  Previously we could consume to a tuple, pair, or expandable
container like a vector, but not to an std::array.  Now we can.

It also enlarges the recognized things in `consume` to allow any
arbitrary deserializable list/dict types (including tuple, and now
array).  You could get previously get them via `consume_list<mytype>()`,
but `consume<mytype>()` didn't accept/recognize it as a list type: now
it does, and similarly for dict-like types.  (This also implicitly
extends to `require<T>`).